Allow to recursively copy a directory in the base docker image (#24)
This PR allows to copy an entire directory into the building tree of the base image.
This was needed by #Deepomatic/vulcain/489 to build protocol buffers into the base image.

The serializers were updated to support a need type of field, i.e. the type "path" (former type "path" was renamed into "file"). So now we have 3 choices for serializers:

"path" : either a file or a directory
"file" : a file
"dir" : a directory
The base image will store a md5 hash of the list of the md5 hashes of the directory, so we won't be able to tell specifically which file has changed and leads to a re-build of the base image. It is OK has it is purely informative.

Also the copy_files field should be renamed in copy_path or something else that suggests that directories are accepted in futur versions.