Don't stop dmake containers on error for dmake tests on jenkins.
Now a slack notification to `#jenkins-dmake` will be sent, containing
a link to the failed job, which is waiting user input to proceed to
the dmake cleanup.

Two actions are proposed: `'Stop all containers' or 'Abort'`; both
with continue and cleanup the dmake run.

Only enabled on the dmake repo test, through
`DMAKE_DEBUG_SHELL_SCRIPTS`.

Also enable `DMAKE_DEBUG` to trace shell scripts on the dmake repo
test.