`kubectl` fails on error only when it returns a non-zero return code
Used and changed semantic for `common.run_shell_command`
`raise_on_return_code`: now it raises an error *only* on non-zero
return code.
Previously it also returned an error if stderr is non-empty, which
makes no sense, but is kept for backward compatibility.

Example:
```
W1205 16:54:23.957959   14961 factory_object_mapping.go:423] Failed to download OpenAPI (the server could not find the requested resource), falling back to swagger
```
`kubectl` writes a warning on its stderr when using client 1.8 with
server 1.6; but it's just a warning, the command worked.