Allow tests without explicit `config` or `config.docker_image`
Default values for `config` and `config.docker_image` are enough to
create a service image that is testable.

We just miss a `config.docker_image.start_script` (generated as `CMD`
in Dockerfile) which is currently interpreted as "the service is not
runnable, nor deployable (which includes publishing the docker
image)).

Implementation details:
- `config` and `config.docker_image` are now not optional anymore, but
since all their children fields are optional or have default values,
it doesn't break anything, and it makes the objects usable, which is
what was needed.
- cleanup: removed all checks for `config.has_value()` and
`docker_image.has_value()` since they always have values
now (sometimes default ones).
- cleanup: removed useless
`DeployConfigSerializer.generate_build_docker` wrapper
- a service without start_script is now non-runnable: it's an error if
it is run (via explicit `dmake run` or via a `needed_service`) or
deployed (which implies running the service).