Fix check_daemon ignored errors: stop using export.
When using export, the command is export, so if the export
succeeded (i.e. always), the return code is zero: `set -e` never
fails, even if the command executed to fill the env var returns a
non-zero status code.
See http://mywiki.wooledge.org/BashFAQ/105 (search for `export`)

It happened when check_daemon failed (for example because the
docker_links container crashes at startup.). The result was ugly: the
env var `DOCKER_LINK_OPS` was filled with stdout of the crashed
container logs, and was passed as argument of the next docker run...

We could fix the export by splitting the command in two:
- variable affectation: if evaluated command fails, `set -e` will be
  triggered
- then exporting the env var

But in fact we don't even need to export the variable: only usages are
DOCKER_LINK_OPS and DAEMON_ID which are used subsequently in the same
shell.

Extra safety: also fix the "env" command, even if it is not used with
evaluated commands but plain values.