Add a Jenkins lock on 'GPU' when testing services that need one
In practice it wraps the whole `Testing App` stage with a
`lock('GPU'){...}` block if any used service (app or docker_link)
requested a gpu via `need_gpu`.

Limitations:
- `dmake shell` is not supported: it is currently split into the
`Deploy` stage instead of `Testing App`, so it's harder to include it
in the lock, and `dmake shell` is never used with the Jenkins pipeline
mode anyway.
- locking is not minimal: we lock the whole testing stage instead of
just when services needing a GPU are up; but it doesn't change much:
we currently have only optimal service start, not service stop, so we
would not gain much. Also, services dependencies are recursive, so we
often need the GPU the whole time anyway.