Add pruning to kubernetes deployment
Previously when removing a resource it leaked in the cluster, because
`kubectl apply` cannot know for sure what to remove.

The way to do it is to explicitly ask for pruning, and limit scope to
some resources via label selectors.
For it to work, we need to
- set a specific label to all resources (top level only because
  cascading handles that (e.g. Deploy -> RC -> Pod)), that
  scopes to the current dmake deployment: dmake app service_name:
  `dmake.deepomatic.com/service=<app>-<service_name>`
  (later it could be changed to deployment, when we add the concept of
  deployment independent of dmake service; the transition may be hard
  though.)
- and select on that when pruning (and when applying: both
  are done at the same time and both are impacted by the label
  selection).

The easiest way is to load the generate yamls, patch them, and
re-dump them, because we have multiple resources:
- env configmap: dmake generates it: we add proper labels directly
- user configmaps from files: dmake generates it but via
  `kubectl create configmap`, which doesn't support adding labels
  yet (see https://github.com/kubernetes/kubernetes/issues/60295):
  we patch the yaml file.
- user manifest: fully under user control: we patch the yaml file.

For easier rollback at kubernetes level, we want to keep old env
configmaps. For that we:
- add a second label: `dmake.deepomatic.com/prune=no-pruning`
  (could add other values later if needed)
  for the dmake generated env configmap, to distinguish them from
  other resources in the same apply scope (dmake service)
- `kubectl apply` in two steps:
  - first for non-pruning manifests (the dmake generated env
    configmap), without `--prune`
  - second for all other manifests, with `--prune` (user
    manifests + user configmaps from files)
(We are forced to do two `kubectl apply` because the label selectors
apply both to the apply and prune parts.)