service docker_image.build: variable substitution for labels in addition to args
Useful to pass custom build args via environment variables.

Also applied "strict" mode on both `args` and `labels`: they are new
options, we should always use "strict" for new options.
"strict" mode raises an error if an undefined variable is used, or if
an error happens during evaluation.

If you want to allow an undefined variable, then use `${MY_VAR:-}`; it
will evaluate to empty string (we have no way to unset...).

Evaluation is done at serialization time: only if the service is
really built, so it should not be too cumbersome to use.