Quickfix: in test mode, always test dependencies before running them
Previously with `dmake test -d '*'` we did not guarantee steps order
between running a service as a dependency of another, and testing that
service.
This resulted in unwanted parallel execution of tests and run as
dependency.

Quickfix: always force testing a service before running it in test
mode (commands `test` and `deploy`).

Drawbacks:
- hardcoded `test` mode, we could use the dependency DAG to know if we
are in test mode instead
- when testing a service instead of `*`, we now also test its
dependencies services.