Add `env_exports` to `needed_services`
This is only used by the parent service when constructing its runtime
environment.
Thus, it does not change how the needed service should be started,
thus we don't need to distinguish it from other instances of
NeededService to build the global commands DAG: just changed `__str__`
and `__repr__`, not `__ne__` nor `__hash__`.