Explain that Jenkins jobs rerun are not supported when it probably happened.
Jenkins (notably the blue ocean interface) proposes everywhere to
rerun the job. Alas it is *not* supported by DMake, because rerun
masks dynamically generated Jenkinsfiles which depend on dynamically
generated temporary directories, which are cleaned at the end of the
first run => rerun always fail, after some steps, with unhelpful
errors.

Now we start the pipeline by checking that the tmp dir exists, and
print a useful error message if not.