Add DMAKE_WITH_DEPENDENCIES option to Jenkins
Useful to execute `dmake test` with no dependencies.

Also, moved the dmake command selection between `deploy` and
`test` (on Jenkins, if in a PR or not) from dmake itself to the
`Jenkinsfile` instead.
It's cleaner, and easier when controlling dependencies (for `deploy`
`dependencies` is hardcoded to true independently of the initial
commandline).