Add OVERRIDE_DMAKE_COMMAND
Related to #377:
(at least when using /rebuild via PR comments) the `params` is
somehow *not* re-evaluated: it is a copy of the previous
execution. This means the Jenkinsfiles *cannot* control the `params`
by setting `DEFAULT_*` global variables.

The workaround consists in using a new set of global variables:
`OVERRIDE_*`, which are evaluated *after* using `params`.