Jenkins Parallel Execution, by internal graph height
This MVP only works on Jenkins, where we have higher-level
primitives, and more resources control.

Enable it via `DMAKE_PARALLEL_EXECUTION = 1` environment variable.

Building blocks:
- use `parallel()` which runs multiple branches of code in parallel,
  in one common executor
  Limitation: need static definition of parallel branches, Jenkins
  doesn't support dynamic generation of branches :-(
- use `lock(PARALLEL_BUILDERS)` to globally limit parallelism on the
  whole Jenkins machine, across jobs, to limit resource
  saturation (mainly RAM)
Locking in a parallel branch doesn't use any (scarce) resource, so the
pattern of `parallel+lock` is OK.

As an MVP, we executes nodes of same height in parallel: the global
execution is a sequence of parallel executions of steps of same
height.
This is not optimal but:
- we cannot reach optimal solution with static scheduling and unknown
  tasks duration,
- it works by construction, nodes depend (directly or indirectly) on
  nodes of strictly lower height, so same height are guaranteed to be
  independent: we can execute them in parallel
- and it gives some notable gains on standard patterns:
  - `base_build` then `build` steps are executed in parallel
  - `needed_links` are started as soon as possible: while the rest
    is built
  - with any luck, some tests are executed in parallel

Later versions could implement some heuristics, maybe detect
independent sub-graphs after having removed base_build and build,
i.e. for run and test, and execute them in parallel (nested
`parallel()`, it works).

From the user point-of-view, we display the standard ordered Plan, by
command, then, as a second pass, display (and dump) the new parallel
plan, by height.

Special case: all `deploy` nodes are executed in parallel, after
everything else (and without resource lock): we don't want to start
deploying some services of an app while other services are still being
tested: their tests could fail and we don't want a partially deployed
app.
Don't lock PARALLEL_BUILDERS on deploy height, it could lead to
deployment deadlock if there is a deployment runtime dependancy
between services.

Implementation-wise, we generate the commands for each node only once:
in the first, default, pass, as before; additionally, we now store
them in a new `nodes_commands[node]` dict.
This is reused in the second pass to re-order by height, and wrap with
`parallel()` and `lock()`.
For the regrouping of `deploy` nodes, we just extract them while
constructing the `nodes_by_height` dict, and inject them back at the
end on a new height.

Special care is taken on GPU lock to start locking it as late as
possible, but since the nodes are now executed as soon as possible, we
may see more GPU contention.
We cannot unlock GPU before the very end of the dmake execution
because there is no notion of `run_stop` for now (we could stop and
unlock gpu before deploying for example)
Caveat: the height on the debug graph is thus wrong in regard to the
parallel execution; it was to much hack to have the graph with
"new/faked" height as in parallel execution while *not* changing
anything to the initial/default Plan.

Visually on Jenkins:

Classic UI:
- logs are prefixed by the parallel branch name
  e.g. `[base @ dmake-test-web-base::base]`
- `Pipeline Steps` show a tree representation, with stage and ~step
  names, durations, and access to log for each step

Blue Ocean:
- the sequence of parallel branches is displayed \o/
- parallel blocs labels are not displayed :-(
- each node is labeled, and clickable to get steps logs
- parallel blocs with zero branches are skipped (otherwise Jenkins
  raises an error)
- parallel blocs with one branch are emitted as such, but create
  graphical glitch on blue ocean with our current jenkins version: the
  node is not visible during the build; but after the job has
  finished, it is visible
  (I tried to emit a stage without `parallel()` in such case, but a
  worse bug happened: the node disappeared not during but after the
  build...)