Don't invalidate base image build cache when fixing largely unused `python_requirements`
1df5824fb7b0a044e1cb894d17ab866a33b55762 / #477 fixed the scripts
behind `python_requirements` (and `python3_requirements`), which were
deprecated a long long time ago (see #77).

Sadly, in doing so, the md5 of the base image source changed, because
these scripts were always added to the base image, even when unused.
This resulted in used visible issue: invalidation of all base images,
for nothing, since this feature is (almost) not used.

The end-goal is to remove that altogether, but in the meantime we can
avoid a double base image hash change by anticipating that end-goal,
while still keeping the feature working for now: don't emit these
files when they are not needed.