Fix uniqueness of NeededService name suffix
Previously, a NeededService was `_specialized` when its runtime was
different than default, but only in the sense of different `env`.
But in fact in the dmake tests we have a service named `test-e2e`
needing 2 versions of `test-external-dependency-nginx`: same default
env, but different link_name.
With previous code these 2 NeededServices had different hash (legit),
so different existences, but they were not _specialized, so they had
no unique suffix distinguishing them; which breaks node string usage
in Jenkinsfile parallel branches.

Remove broken notion of specialized NeededService; just do the ad-hoc
check in the only location where it was used: generating a unique
suffix for non default NeededServices.

Tricky part: we probably do need to distinguish 2 NeededServices with
different env_exports (so different hash), but to be conservative we
don't want to change the uniquified needed service name; maybe we
should?