Add needed_services[].use_host_ports
Useful to force disable host_port when a service is used as a
dependency, and still keep this available when that service is
directly started; for example a frontend, also started as a dependency
of the backend, to support front, back, and both-together development
modes.

Add useless host_port and use_host_ports usage in test/*/dmake.yml to
cover that part.

This doesn't properly cover all edge cases regarding NeededService
specialization: we assume all instances for a service that are
otherwise identical are using the same value for
`use_host_ports`. It's undefined behavior to not do that.

Also improve test/*/.dockerignore for faster tests.