Fix cobertura_report collection error because of earlier error, before tests
CoberturaPublisher is done at the end of the execution because it
cannot be duplicated (at least it was the case several years ago when
it was plugged-in).

With 1f3c17db399c85e923e5506e1c3d33636473bb57 / #482 we tried to
eagerly collect tests results even when tests failed, but didn't
properly handle that late collection in the case of errors before any
test step: the tests results won't exist.

Because it's a global, late collection, we cannot do the fancy double
collection with/without error propagation that we did in #482 for the
other tests results collections.
That's OK; the situation is explained to the user.