kubernetes annotations: set both real git branch and dmake environment branch name
closes #520

When deploying on virtual-git-branches-as-dmake-environments, we select the env with `--branch`.
This previously resulted in the `dmake.deepomatic.com/git-branch` annotation containing the environment name instead of the real underlying git branch name, which is more informative of what code is really deployed.

We now generate both values, as well as print them in the header logs
when they are different.