Add new dmake arg for Jenkinsfile control: dmake_arg_deploy_region
When set (from a new Jenkinsfile.<deploy_region>), dmake is called
with `--branch=<the_actual_gitbranch>-<deploy_region>`, using the
'environment branch' suffix as a ~deployment profile for different
regions.

dmake Jenkinsfile entrypoint previously supported:
- `DEFAULT_DMAKE_*` groovy variables to change the default job
  parameters values
- `OVERRIDE_DMAKE_*` groovy variables to override parameters (hackish)

This introduces a new kind:
- `dmake_arg_*` groovy variables to pass dmake arguments which are not
  exposed as job parameters

This is not (yet) supported for Jenkins PR builds as we don't need it,
and extra care would need to be taken notably for the kubernetes
manifest pre-run test.

This is easy when limited to deployment branch jobs: it's for
`env.BRANCH_NAME`.